cleanup to make typescript/eslint play nice
